Unit tests cluster_manager.py

+++ simulation-system/libs/csle-collector/src/csle_collector/snort_ids_manager/dao/snort_ids_ip_alert_counters.py
@@ -14,8 +14,9 @@ class SnortIdsIPAlertCounters:
         """
         Initializes the DTO
         """
-        self.priority_alerts = list(np.zeros(4).tolist())
-        self.class_alerts = list(np.zeros(len(set(constants.SNORT_IDS_ROUTER.SNORT_ALERT_IDS_ID.values()))).tolist())
+        self.priority_alerts = list(np.zeros(4).astype(int).tolist())
+        self.class_alerts = list(np.zeros(len(set(
+            constants.SNORT_IDS_ROUTER.SNORT_ALERT_IDS_ID.values()))).astype(int).tolist())
         self.severe_alerts = 0
         self.warning_alerts = 0
         self.total_alerts = 0
